refactor(org): batch fetch child group entities

chore: add types to OwnershipCard
