+++ plugins/org/src/components/Cards/OwnershipCard/types.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Backstage Authors
+ * Copyright 2023 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,5 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from './OwnershipCard';
-export * from './types';
+
+/** @public */
+export const DefaultRelationType = {
+  Direct: 'direct',
+  Aggregated: 'aggregated',
+} as const;
+
+/** @public */
+export type RelationType =
+  (typeof DefaultRelationType)[keyof typeof DefaultRelationType];
